feat: store auth_token between different calls
currently /api/2.26/arrays/space is called every 10 seconds from each proxmox node.
Each of these calls is creating a new auth token on the storage because the token
in stored in $scfg is not persistent.

This patch adds a mechanism to store the token in an external file between differnt
calls to avoid new session creation at (almost) each api call.

Signed-off-by: Peter Lieven <pl@dlhnet.de>